backend/requirements.txt
fastapi
uvicorn
pymongo
pandas
google-genai
python-multipart
python-dotenv
openpyxl
pydantic
fpdf2
matplotlib
